Fix capsule route registration order.
